core/LuaTouchpadGesture: split out code from header
